chore: move styles to page components
