+++ client/src/pages/p/[id].tsx
@@ -11,7 +11,29 @@ import { AppLoader } from "../../components/Loader";
 import { localeInfo } from "../../config";
 import useRandomGPromise from "../../hooks/useRandomGPromise";
 import { GPromise } from "../../components/GPromise";
-import { ForwardIcon, ShareButton, ShareIcon, TwitterIcon } from "./styles";
+import { FaForward, FaShareAlt, FaTwitter } from "react-icons/fa";
+import styled, { css } from "styled-components";
+
+export const ShareButton = styled.button`
+  background-color: transparent;
+  border: none;
+  color: ${({ theme }) => theme.color.brand};
+  cursor: pointer;
+`;
+
+const iconStyle = css``;
+
+export const TwitterIcon = styled(FaTwitter)`
+  ${iconStyle}
+`;
+
+export const ShareIcon = styled(FaShareAlt)`
+  ${iconStyle}
+`;
+
+export const ForwardIcon = styled(FaForward)`
+  ${iconStyle}
+`;
 
 export const getServerSideProps = async (
   context: GetServerSidePropsContext
